+++ lib/src/calculator_text_field.dart
@@ -20,6 +20,7 @@ class CalculatorTextField extends StatefulWidget with BaseTextField {
     this.style,
     this.strutStyle,
     this.textAlign = TextAlign.start,
+    this.valueFormat,
   }) : super(key: key);
 
   final String title;
@@ -37,6 +38,7 @@ class CalculatorTextField extends StatefulWidget with BaseTextField {
   final TextStyle style;
   final StrutStyle strutStyle;
   final TextAlign textAlign;
+  final ValueFormat<double> valueFormat;
 
   @override
   _CalculatorTextFieldState createState() => _CalculatorTextFieldState();
@@ -44,6 +46,18 @@ class CalculatorTextField extends StatefulWidget with BaseTextField {
 
 class _CalculatorTextFieldState extends State<CalculatorTextField> {
   final inputController = TextEditingController();
+
+  @override
+  void initState() {
+    super.initState();
+
+    setValue(widget.initialValue);
+  }
+
+  void setValue(value) {
+    inputController.text =
+        widget.valueFormat != null ? widget.valueFormat(value) : '$value';
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -55,11 +69,11 @@ class _CalculatorTextFieldState extends State<CalculatorTextField> {
       textAlign: widget.textAlign,
       decoration: widget.inputDecoration,
       onTap: () async {
-        final resutl = await widget.showInputCalculator(context);
+        final result = await widget.showInputCalculator(context);
         setState(() {
-          inputController.text = '${resutl ?? 0.0}';
+          setValue(result);
 
-          if (widget.onSubmitted != null) widget.onSubmitted(resutl);
+          if (widget.onSubmitted != null) widget.onSubmitted(result);
         });
       },
     );